+++ src/pages/SchoolDetail/TabsDetail/TabsD.module.css
@@ -1,0 +1,31 @@
+.divBox{
+    padding-top: 8vh;
+    width: 120vh;
+    min-height: 50vh;
+  }
+
+@media (max-width: 900px) {
+    .divSwipperCitas{
+      /* width:'100%',display:'flex', justifyContent:'center' */
+      width: 100%;
+      display: flex;
+      justify-content: center;
+    }
+    .divBox{
+      width: auto;
+    }
+    .responsiveHead{
+      display: flex;
+      flex-direction: column;
+      justify-content: flex-start;
+      align-items: flex-start;
+     padding: 2vh;
+    }
+    .divIconsHead{
+    padding-left: 2vh;
+  
+  
+      
+  
+    }
+  }
